requirements.txt
# Install with: pip install -r requirements.txt
# Then: python -m playwright install chromium

fastmcp>=2.0
playwright>=1.40
pydantic-settings>=2.0
structlog>=24.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21
